requirements.txt
PyPDF2
spacy
nltk
pdf2image
Pillow
openvino
fitz
PyMuPDF
